Libplanet 0.10.0-dev.20200611035053로 올리고 NCG를 fungible asset states로 재구현

+++ nekoyume/Assets/_Scripts/Lib9c/Model/Quest/Quest.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using Bencodex.Types;
 using Nekoyume.Model.EnumType;
 using Nekoyume.Model.Item;
@@ -262,7 +263,7 @@ namespace Nekoyume.Model.Quest
             }
         }
 
-        public void UpdateTradeQuest(TradeType type, decimal price)
+        public void UpdateTradeQuest(TradeType type, BigInteger price)
         {
             var tradeQuests = _quests
                 .OfType<TradeQuest>()
