+++ nekoyume/Assets/_Scripts/UI/Model/ShopItem.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Numerics;
 using Libplanet;
 using Nekoyume.Model.Item;
 using Nekoyume.UI.Module;
@@ -10,17 +11,17 @@ namespace Nekoyume.UI.Model
     {
         public readonly ReactiveProperty<Address> SellerAgentAddress = new ReactiveProperty<Address>();
         public readonly ReactiveProperty<Address> SellerAvatarAddress = new ReactiveProperty<Address>();
-        public readonly ReactiveProperty<decimal> Price = new ReactiveProperty<decimal>();
+        public readonly ReactiveProperty<BigInteger> Price = new ReactiveProperty<BigInteger>();
         public readonly ReactiveProperty<Guid> ProductId = new ReactiveProperty<Guid>();
-        
+
         public ShopItemView View;
-        
+
         public ShopItem(Address sellerAgentAddress, Nekoyume.Model.Item.ShopItem item)
             : this(sellerAgentAddress, item.SellerAvatarAddress, item.Price, item.ProductId, item.ItemUsable)
         {
         }
 
-        private ShopItem(Address sellerAgentAddress, Address sellerAvatarAddress, decimal price, Guid productId,
+        private ShopItem(Address sellerAgentAddress, Address sellerAvatarAddress, BigInteger price, Guid productId,
             ItemBase item) : base(item, 1)
         {
             GradeEnabled.Value = true;
